Move out fixtures from test_run_dialog
